pyproject.toml
[project]
name = "financial-report-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.14.0",
    "llama-cloud-services>=0.6.21",
    "openai>=1.76.0",
    "pikepdf>=9.7.0",
    "pymupdf>=1.25.5",
]
